Made uplaoding an image save to the db
